Put MoveOrderer and MoveScorer in separate files.
