+++ src/chess-engine-lib/CMakeLists.txt
@@ -11,7 +11,8 @@ set(
     "GameState.cpp"
     "GameStateStringOps.cpp"
     "Move.cpp"
-    "MoveOrdering.cpp"
+    "MoveOrderer.cpp"
+    "MoveScorer.cpp"
     "MoveSearcher.cpp"
     "PawnMasks.cpp"
     "Perft.cpp"
